Configure server to serve the index.html file regardless of url
In order for React Router to handle routing after the page refreshes, the server must send the index.html file to the client for every route. If the server does not handle unknown routes like this,
React Router will not have a chance to evaluate the route and render the correct components as the server will respont with a 404 error.