Pass setReviewList and characteristics to subcomponents
Reviews now has access to the setRevewList action dispatcher and passes it to the ReviewList component so that it can update the list of reviews to display when a button inside of it is clicked.
RatingBreakdown will eventually display the characteristics of a product, so characteristics is now being passed to its props.