Create CharacteristicForm component
The CharacteristicForm accepts a name (string) and a handler (function). It renders out a form depending on the characteristic name (name parameter) that allows the user to select a
characteristic rating. The handler function is run with the currently selected value passed in as an argument whenever the value of the radio button group changes.