Implement re-rendering when reviews are changed and when helpfulness is increased
Previously when clicking the button to increase helpfulness in the Review Item component, the component would not detect the change and re-render to display the new helpfulness count.
That is because the hook for handling state and causing renders was not checking if the reviews changed. However, since the "reviews" argument is an object, it cannot be passed into the hook's
arguments for values to be watched. The "reviewsChanged" state variable allows renders to be invoked when changing it's value.

A state variable for keeping track of reviews that have already been marked helpful (helpfulReviews) has also been implemented so a user must refresh to mark a review helpful again.