Implement button for incrementing the helpful count on a review
When "yes" is clicked, the review is marked helpful and the list of reviews is rendered again to show this change. The user is then unable to mark the review as helpful again until they refresh the page.
the "report" button has also been started but is currently not functional.