Implement "report" functionality on individual reviews
When the "report" item is clicked in a review component, an API call is made with a request to report (and subsequently delete) that review. After the review is reported, the now reported review
is replaced with a placeholder component that indicates to the user that the review has been reported.