Implement overflow handling for individual reviews.
When a review has body text that is greater than 250 characters, the text is cut off and a "show more" link appears. When it is clicked, the full text of the review is revealed. Also, review
body text is now broken up to prevent long reviews from changing container width.