Change report functionality on reviews
Previously when reporting a review, the whole review would be removed immediately and replaced with text indicating that the review was reported. Now the review remains but the "Report" button
changes to "Reported" without underlining or changing your cursor to a pointer, and the review is then removed on page re-render.